requirements.txt
tiktoken
streamlit
elasticsearch<9,>=8.12
pypdf==4.2.0
python-pptx>=1.0.2
pdfplumber==0.11.4
Pillow==10.4.0
nltk==3.9.1
python-slugify==8.0.4
python-dotenv==1.0.1
python-docx>=1.1.0
ibm-watson
ibm-cloud-sdk-core
docling==2.58.0
faster-whisper==1.0.3
requests>=2.31
